runtime/requirements.txt
tushare>=1.2.89
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
markdown2>=2.4.10
requests>=2.28.0
openpyxl>=3.1.0
pdfplumber>=0.11.0
PyYAML>=6.0.0
